[http] Pull StateRequest and CommandRequest out as top-level classes
Also remove custom HTTP method type in favor of the one provided by
Jetty.
